Move upload process to its own file
